requirements.txt
python-consul
pyyaml
